server/requirements.txt
fastapi
pydantic
uvicorn
aiohttp
python-dotenv
python-dateutil
pytz
undetected-chromedriver
selenium
supabase
openai
